@@ -8,8 +8,6 @@
 		"test": "ng test",
 		"lint": "ng lint",
 		"e2e": "ng e2e",
-		"format:fix": "pretty-quick --staged --config ./.prettierrc ./src",
-		"precommit": "run-s format:fix lint",
 		"storybook": "start-storybook -p 6006",
 		"build-storybook": "build-storybook"
 	},
@@ -25,15 +23,18 @@
 		"@angular/platform-browser-dynamic": "^6.1.0",
 		"@angular/router": "^6.1.0",
 		"@ngrx/effects": "^6.1.0",
+		"@ngrx/router-store": "^6.1.0",
 		"@ngrx/store": "^6.1.0",
 		"@ngrx/store-devtools": "^6.1.0",
 		"@types/query-string": "^6.1.0",
 		"core-js": "^2.5.4",
 		"fast-deep-equal": "^2.0.1",
+		"lodash": "^4.17.10",
 		"moment": "^2.22.2",
+		"ng2-dragula": "^2.0.2",
 		"normalizr": "^3.2.4",
-		"primeng": "^6.1.0-rc.1",
 		"primeicons": "^1.0.0-beta.10",
+		"primeng": "^6.1.0-rc.1",
 		"query-string": "^6.1.0",
 		"rxjs": "^6.0.0",
 		"uuid": "^3.3.2",
@@ -55,7 +56,6 @@
 		"@types/node": "~8.9.4",
 		"babel-core": "^6.26.3",
 		"codelyzer": "~4.2.1",
-		"husky": "^0.14.3",
 		"jasmine-core": "~2.99.1",
 		"jasmine-spec-reporter": "~4.2.1",
 		"karma": "~1.7.1",
@@ -63,12 +63,9 @@
 		"karma-coverage-istanbul-reporter": "~2.0.0",
 		"karma-jasmine": "~1.1.1",
 		"karma-jasmine-html-reporter": "^0.2.2",
-		"npm-run-all": "^4.1.3",
-		"prettier": "^1.14.0",
-		"pretty-quick": "^1.6.0",
 		"protractor": "~5.3.0",
 		"ts-node": "~5.0.1",
 		"tslint": "~5.9.1",
-		"typescript": "~2.7.2"
+		"typescript": "^2.9.2"
 	}
 }
