@@ -1,9 +1,10 @@
 import { NgModule } from '@angular/core';
 import { HttpClientModule } from '@angular/common/http';
 import { HttpService } from '@app/api/http/http.service';
+import { UserDomainService } from '@app/api/domains/user/user.domain';
 
 @NgModule({
 	imports: [HttpClientModule],
-	providers: [HttpService]
+	providers: [HttpService, UserDomainService]
 })
 export class APIModule {}
