@@ -1,12 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 
+
 @Component({
 	selector: 'app-app',
 	templateUrl: 'app.component.html',
 	styleUrls: ['./app.component.sass']
 })
 export class AppComponent implements OnInit {
-	constructor() {}
-
 	ngOnInit() {}
 }
