@@ -9,9 +9,23 @@ import { CustomizeChartComponent } from '@app/features/draftproject/custom-chart
 import { ChartComponent } from '@app/features/draftproject/custom-chart/chart/chart.component';
 import { ExportComponent } from '@app/features/draftproject/export/export.component';
 import { ChartPreviewComponent } from '@app/features/draftproject/list-charts/chart-preview/chart-preview.component';
+import { ProjectNameComponent } from '@app/features/draftproject/project-name/project-name.component';
+import { CommonModule } from '@angular/common';
+import { BrowserModule } from '@angular/platform-browser';
+import { FormsModule } from '@angular/forms';
+import { SharedModule } from '@app/shared/shared.module';
+import { TabViewModule } from 'primeng/tabview';
+import { FileUploadModule } from 'primeng/fileupload';
 
 @NgModule({
-	imports: [],
+	imports: [
+		SharedModule,
+		TabViewModule,
+		FileUploadModule,
+		CommonModule,
+		BrowserModule,
+		FormsModule
+	],
 	declarations: [
 		StepperComponent,
 		LoadDataComponent,
@@ -22,7 +36,8 @@ import { ChartPreviewComponent } from '@app/features/draftproject/list-charts/ch
 		CustomizeChartComponent,
 		ChartComponent,
 		ExportComponent,
-		ChartPreviewComponent
+		ChartPreviewComponent,
+		ProjectNameComponent
 	],
 	exports: [
 		StepperComponent,
@@ -34,7 +49,8 @@ import { ChartPreviewComponent } from '@app/features/draftproject/list-charts/ch
 		CustomizeChartComponent,
 		ChartComponent,
 		ExportComponent,
-		ChartPreviewComponent
+		ChartPreviewComponent,
+		ProjectNameComponent
 	]
 })
 export class DraftprojectModule {}
