@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { DraftprojectModule } from '@app/features/draftproject/draftproject.module';
+import { AuthenticationModule } from '@app/features/authentication/authentication.module';
 
 @NgModule({
-	imports: [DraftprojectModule],
-	exports: [DraftprojectModule]
+	imports: [DraftprojectModule, AuthenticationModule],
+	exports: [DraftprojectModule, AuthenticationModule]
 })
 export class FeatureModule {}
