@@ -1,5 +1,5 @@
 export interface Company {
-	id?: string;
-	name?: string;
-	createdAt?: string | number;
+	id: string;
+	name: string;
+	createdAt: string | number;
 }
