@@ -1,8 +1,8 @@
-import { AppAction } from './store.model';
+import { AppAction } from '@app/models/store.model';
 
 export interface ErrorHandlerState {
-	action?: AppAction;
-	msg?: string;
+	action: AppAction;
+	msg: string;
 	serverMessages?: string[];
 	error?: Error;
 	isOpen?: boolean;
