@@ -1,8 +1,3 @@
-export class LoginForm {
-	constructor(
-		public email: string = null,
-		public password: string = null,
-		public lastName: string = null,
-		public firstName: string = null
-	) {}
+export class Login {
+	constructor(public email: string = null, public password: string = null) {}
 }
