@@ -1,7 +1,12 @@
-import { Project } from './project.model';
+import { Project } from '@app/models/project.model';
+import {
+	NormalizedSchemeWithFetching,
+	SchemeID,
+	NormalizedActiveEntity
+} from '@app/models/normalizr.model';
 
-export interface ProjectsState {
-	all: string[];
-	byId: { [id: string]: Project };
-	isLoading: boolean;
+export class ProjectsState<U = SchemeID, R = undefined>
+	extends NormalizedSchemeWithFetching<Project, R>
+	implements NormalizedActiveEntity<U, R> {
+	active = null;
 }
