@@ -1,11 +1,12 @@
-import { HttpHeaders } from '@angular/common/http';
 import { RequestType } from '@app/models/requestType.model';
 
 export class ServiceRequest {
 	constructor(
 		public readonly type: RequestType,
 		public readonly url: string,
-		public readonly headers?: HttpHeaders,
+		public readonly headers?: {
+			[key: string]: string;
+		},
 		public readonly payload?: any
 	) {}
 }
