@@ -1,9 +1,12 @@
+import { SchemeID } from '@app/models/normalizr.model';
+
 export interface User {
-	id?: string | number;
-	name?: string;
-	email?: string;
-	avatar?: string;
-	firstName?: string;
-	lastName?: string;
-	createdAt?: number | string;
+	id: SchemeID;
+	name: string;
+	email: string;
+}
+
+export interface AuthenticatedUser {
+	token: string;
+	user: User;
 }
