@@ -1,5 +1,9 @@
 import { Injectable } from '@angular/core';
-import { FormControl, AbstractControlOptions } from '@angular/forms';
+import {
+	FormControl,
+	AbstractControlOptions,
+	AbstractControl
+} from '@angular/forms';
 import { fieldsValidators, OptionalType } from '@app/models';
 
 @Injectable()
@@ -9,7 +13,7 @@ export class FormService {
 	createFormControls<T>(
 		initialValues: OptionalType<T>,
 		validators: fieldsValidators<OptionalType<T>>
-	) {
+	): { [key: string]: AbstractControl } {
 		return Object.keys(initialValues).reduce((previous, key) => {
 			previous[key] = new FormControl(initialValues[key], validators[
 				key
