@@ -3,7 +3,9 @@ import { CommonModule } from '@angular/common';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { BrowserModule } from '@angular/platform-browser';
 import { TabMenuModule } from 'primeng/tabmenu';
+
 // Forms
+import { FormsModule } from '@angular/forms';
 import { PasswordModule } from 'primeng/password';
 import { ButtonModule } from 'primeng/button';
 import { RadioButtonModule } from 'primeng/radiobutton';
@@ -17,6 +19,10 @@ import { AutoCompleteModule } from 'primeng/autocomplete';
 import { InputSwitchModule } from 'primeng/inputswitch';
 import { ToolbarModule } from 'primeng/toolbar';
 import { TabViewModule } from 'primeng/tabview';
+
+// Drag&Drop
+import { DragulaModule } from 'ng2-dragula';
+import { TooltipModule } from 'primeng/tooltip';
 
 @NgModule({
 	imports: [
@@ -36,7 +42,10 @@ import { TabViewModule } from 'primeng/tabview';
 		InputSwitchModule,
 		ToolbarModule,
 		TabMenuModule,
-		TabViewModule
+		TabViewModule,
+		DragulaModule.forRoot(),
+		FormsModule,
+		TooltipModule
 	],
 	exports: [
 		ButtonModule,
@@ -55,7 +64,10 @@ import { TabViewModule } from 'primeng/tabview';
 		InputSwitchModule,
 		ToolbarModule,
 		TabMenuModule,
-		TabViewModule
+		TabViewModule,
+		DragulaModule,
+		FormsModule,
+		TooltipModule
 	],
 	declarations: []
 })
