@@ -1,5 +1,5 @@
 export enum CompaniesActionConstants {
-	COMPANIES_LOAD_DATA = 'Companies: LOAD_DATA',
-	COMPANIES_LOAD_DATA__COMPLETE = 'Companies: LOAD_DATA__COMPLETE',
-	COMPANIES_LOAD_DATA__FAILED = 'Companies: LOAD_DATA__FAILED'
+	LOAD_COMPANIES = 'Companies: LOAD',
+	LOAD_COMPANIES__COMPLETE = 'Companies: LOAD__COMPLETE',
+	LOAD_COMPANIES__FAILED = 'Companies: LOAD__FAILED'
 }
