@@ -1,5 +1,11 @@
 export enum ProjectsActionConstants {
-	PROJECTS_LOAD_DATA = 'Projects: LOAD_DATA',
-	PROJECTS_LOAD_DATA__COMPLETE = 'Projects: LOAD_DATA__COMPLETE',
-	PROJECTS_LOAD_DATA__FAILED = 'Projects: LOAD_DATA__FAILED'
+	LOAD_PROJECTS = 'Projects: LOAD',
+	LOAD_PROJECTS__COMPLETE = 'Projects: LOAD__COMPLETE',
+	LOAD_PROJECTS__FAILED = 'Projects: LOAD__FAILED',
+
+	CREATE_DRAFT_PROJECT = 'Project: CREATE_DRAFT',
+	CREATE_DRAFT_PROJECT__COMPLETE = 'Project: CREATE_DRAFT__COMPLETE',
+	CREATE_DRAFT_PROJECT__FAILED = 'Project: CREATE_DRAFT__FAILED',
+
+	CHANGE_PROJECT_NAME = 'Project: CHANGE_NAME'
 }
