@@ -1,5 +1,13 @@
 export enum UserActionConstants {
-	USER_VERIFY_TOKEN = 'User: VERIFY_TOKEN',
-	USER_VERIFY_TOKEN__COMPLETE = 'User: VERIFY_TOKEN__COMPLETE',
-	USER_VERIFY_TOKEN__FAILED = 'User: VERIFY_TOKEN__FAILED'
+	VERIFY_USER_TOKEN = 'User: VERIFY_TOKEN',
+	VERIFY_USER_TOKEN__COMPLETE = 'User: VERIFY_TOKEN__COMPLETE',
+	VERIFY_USER_TOKEN__FAILED = 'User: VERIFY_TOKEN__FAILED',
+
+	LOGIN = 'User: LOGIN',
+	LOGIN__COMPLETE = 'User: LOGIN__COMPLETE',
+	LOGIN__FAILED = 'User: LOGIN__FAILED',
+
+	REGISTER = 'User: REGISTER',
+	REGISTER__COMPLETE = 'User: REGISTER__COMPLETE',
+	REGISTER__FAILED = 'User: REGISTER__FAILED'
 }
