@@ -3,10 +3,18 @@ import { EffectsModule } from '@ngrx/effects';
 import { UserEffects } from '@app/store/effects/user.effects';
 import { CompaniesEffects } from '@app/store/effects/companies.effects';
 import { ProjectsEffects } from '@app/store/effects/projects.effects';
+import { ChartsEffects } from '@app/store/effects/charts.effects';
+import { RouterEffects } from '@app/store/effects/router.effects';
 
 @NgModule({
 	imports: [
-		EffectsModule.forRoot([UserEffects, CompaniesEffects, ProjectsEffects])
+		EffectsModule.forRoot([
+			UserEffects,
+			CompaniesEffects,
+			ChartsEffects,
+			ProjectsEffects,
+			RouterEffects
+		])
 	]
 })
 export class EffectModule {}
