@@ -1,5 +1,5 @@
-import { Actions as UserActions } from '../actions/user/user.actions';
-import { UserActionConstants } from '../actions/user/user.action-types';
+import { Actions as UserActions } from '@app/store/actions/user/user.actions';
+import { UserActionConstants } from '@app/store/actions/user/user.action-types';
 import { User, UserState } from '@app/models';
 import { combineReducers } from '@ngrx/store';
 
@@ -7,21 +7,19 @@ export const initialState: UserState = {
 	info: {
 		id: null,
 		name: null,
-		email: null,
-		avatar: null,
-		firstName: null,
-		lastName: null,
-		createdAt: null
+		email: null
 	},
 	isLoading: null
 };
 
 export const info = (state = initialState.info, action: UserActions): User => {
 	switch (action.type) {
-		case UserActionConstants.USER_VERIFY_TOKEN__COMPLETE: {
+		case UserActionConstants.VERIFY_USER_TOKEN__COMPLETE:
+		case UserActionConstants.LOGIN__COMPLETE:
+		case UserActionConstants.REGISTER__COMPLETE: {
 			return {
 				...state,
-				...action.payload
+				...action.payload.user
 			};
 		}
 		default:
@@ -34,13 +32,19 @@ export const isLoading = (
 	action: UserActions
 ): boolean => {
 	switch (action.type) {
-		case UserActionConstants.USER_VERIFY_TOKEN: {
+		case UserActionConstants.VERIFY_USER_TOKEN:
+		case UserActionConstants.LOGIN:
+		case UserActionConstants.REGISTER:
 			return true;
-		}
-		case UserActionConstants.USER_VERIFY_TOKEN__COMPLETE:
-		case UserActionConstants.USER_VERIFY_TOKEN__FAILED: {
+
+		case UserActionConstants.VERIFY_USER_TOKEN__COMPLETE:
+		case UserActionConstants.VERIFY_USER_TOKEN__FAILED:
+		case UserActionConstants.LOGIN__COMPLETE:
+		case UserActionConstants.LOGIN__FAILED:
+		case UserActionConstants.REGISTER__COMPLETE:
+		case UserActionConstants.REGISTER__FAILED:
 			return false;
-		}
+
 		default:
 			return state;
 	}
