@@ -1,5 +1,4 @@
 import { AppState } from '@app/models';
 
-export const companies = (state: AppState) => {
-	return state.companies.all.map(id => state.companies.byId[id]);
-};
+export const companies = (state: AppState) =>
+	state.companies.all.map(id => state.companies.byId[id]);
