@@ -1,5 +1,5 @@
 import { AppState } from '@app/models';
 
-export const user = (state: AppState) => {
-	return state.user.info;
-};
+export const user = () => (state: AppState) => state.user.info;
+
+export const isUserLoading = () => (state: AppState) => state.user.isLoading;
