@@ -23,7 +23,10 @@ storiesOf('Checkbox', module)
 			props: {
 				label: 'I definetly want to choose it',
 				value: 'first',
-				control: new FormControl('', Validators.required)
+				control: new FormControl(
+					{ value: false, disabled: false },
+					Validators.required
+				)
 			}
 		}),
 		{
@@ -39,7 +42,11 @@ storiesOf('Checkbox', module)
 			props: {
 				label: 'I choose it!',
 				value: 'second',
-				control: new FormControl('', Validators.required)
+				checkedValues: ['second'],
+				control: new FormControl(
+					{ value: true, disabled: false },
+					Validators.required
+				)
 			}
 		}),
 		{
@@ -55,8 +62,10 @@ storiesOf('Checkbox', module)
 			props: {
 				label: 'I`m disabled (',
 				value: 'third',
-				disabled: true,
-				control: new FormControl(true)
+				control: new FormControl(
+					{ value: false, disabled: true },
+					Validators.required
+				)
 			}
 		}),
 		{
@@ -72,9 +81,7 @@ storiesOf('Checkbox', module)
 			props: {
 				label: 'I`m disabled too (',
 				value: 'fifth',
-				disabled: true,
-				selectedValues: ['fifth'],
-				control: new FormControl(true)
+				control: new FormControl({ value: true, disabled: true })
 			}
 		}),
 		{
@@ -92,23 +99,28 @@ storiesOf('Checkbox', module)
 					{
 						label: 'First item',
 						value: 'first',
-						disabled: false,
-						control: new FormControl()
+						control: new FormControl(
+							{ value: false, disabled: false },
+							Validators.required
+						)
 					},
 					{
 						label: 'Seventh item',
 						value: 'second',
-						disabled: false,
-						control: new FormControl()
+						control: new FormControl(
+							{ value: true, disabled: false },
+							Validators.required
+						)
 					},
 					{
 						label: 'Fifth item',
 						value: 'third',
-						disabled: true,
-						control: new FormControl()
+						control: new FormControl(
+							{ value: true, disabled: true },
+							Validators.required
+						)
 					}
-				],
-				selectedValues: ['second', 'third']
+				]
 			}
 		}),
 		{
