@@ -4,11 +4,13 @@ const download = require('download-file');
 
 class FileMiddleware {
 	save(file) {
-		let path = `../server/app/fileStorage/${uuidv4()}${file.name}`;
+		this.createStorageFolder();
+		const storagePath = process.env.FILE_STORAGE_DIR;
+		let path = `${storagePath}${uuidv4()}${file.name}`;
 		return new Promise((resolve, reject) => {
 			// rename file if exists, otherwise it will be overwritten
 			while (fs.existsSync(path)) {
-				path = `../server/app/fileStorage/${uuidv4()}${file.name}`;
+				path = `${storagePath}${uuidv4()}${file.name}`;
 			}
 			const stream = fs.createWriteStream(path);
 			return stream.once('open', () => {
@@ -27,21 +29,20 @@ class FileMiddleware {
 	}
 
 	saveFromLink(url) {
-		this.url = url;
-		return new Promise(resolve => {
+		this.createStorageFolder();
+		const storagePath = process.env.FILE_STORAGE_DIR;
+		return new Promise((resolve, reject) => {
 			const options = {
-				directory: '../server/app/fileStorage/',
+				directory: `${storagePath}`,
 				filename: `${uuidv4()}.file`
 			};
 			while (fs.existsSync(options.directory + options.filename)) {
 				options.filename = `${uuidv4()}.file`;
 			}
-			download(this.url, options, err => {
-				if (err) throw err;
+			download(url, options, err => {
+				if (err) return reject(err);
 				return resolve(options.directory + options.filename);
 			});
-			// todo: fix error handling if its possible
-			// return reject(new Error('Something went wrong'));
 		});
 	}
 
@@ -57,6 +58,13 @@ class FileMiddleware {
 				resolve("File already doesn't exist");
 			}
 		});
+	}
+
+	createStorageFolder() {
+		this.path = process.env.FILE_STORAGE_DIR;
+		if (!fs.existsSync(this.path)) {
+			fs.mkdirSync(this.path);
+		}
 	}
 }
 module.exports = new FileMiddleware();
