@@ -1,12 +1,13 @@
-require('dotenv').config();
 const jwt = require('jsonwebtoken');
+require('dotenv').config();
 
 class TokenService {
 	constructor() {
-		this.tokenSecret = process.env.TOKEN;
+		this.tokenSecret = null;
 	}
 
 	createToken(data) {
+		this.tokenSecret = process.env.TOKEN;
 		const tokenPayload = {
 			id: data.id,
 			firstName: data.firstName,
@@ -14,11 +15,12 @@ class TokenService {
 			email: data.email
 		};
 		return jwt.sign(tokenPayload, this.tokenSecret, {
-			expiresIn: 60 * 60
+			expiresIn: process.env.TOKEN_TIME_EXP
 		});
 	}
 
 	verifyToken(token) {
+		this.tokenSecret = process.env.TOKEN;
 		return new Promise((resolve, reject) => {
 			jwt.verify(token, this.tokenSecret, (err, decoded) => {
 				if (err) {
