@@ -1,11 +1,12 @@
+require('dotenv').config();
 const Sequelize = require('sequelize');
 
 const config = {
-	user: 'postgres',
-	host: 'localhost',
-	database: 'nchart',
-	password: 'admin',
-	port: 5432
+	user: process.env.DB_USER,
+	host: process.env.DB_HOST,
+	database: process.env.DB_NAME,
+	password: process.env.DB_PASS,
+	port: process.env.DB_PORT
 };
 
 module.exports = new Sequelize(
