@@ -1,19 +1,32 @@
-// const Sequelize = require("sequelize");
-// const sequelize = require("../../config/index");
-//
-// const Project = sequelize.define("project", {
-//   group_id: {
-//     type: Sequelize.INTEGER
-//   },
-//   name: {
-//     type: Sequelize.STRING
-//   },
-//   data: {
-//     type: Sequelize.JSON
-//   }
-// });
-//
-// // this method creates table if it doesn't exit
-// Project.sync();
-//
-// module.exports = Project;
+const Sequelize = require('sequelize');
+const sequelize = require('../../../config/index');
+const ProjectChart = require('./project_chart');
+const GroupProject = require('../../group/group.models/group_project');
+
+const Project = sequelize.define('project', {
+	name: {
+		type: Sequelize.STRING,
+		allowNull: false
+	}
+});
+
+Project.sync().then(() => {
+	ProjectChart.sync().then(() => Project.hasMany(ProjectChart, {
+		foreignKey: 'projectId',
+		sourceKey: 'id',
+		onDelete: 'CASCADE',
+		constraints: false
+	}));
+	ProjectChart.belongsTo(Project, { foreignKey: 'projectId' });
+	GroupProject.sync().then(() => {
+		Project.hasMany(GroupProject, {
+			foreignKey: 'projectId',
+			sourceKey: 'id',
+			onDelete: 'CASCADE',
+			constraints: false
+		});
+	});
+	GroupProject.belongsTo(Project, { foreignKey: 'projectId' });
+});
+
+module.exports = Project;
