@@ -1,48 +1,11 @@
 const project = require('express').Router();
-const projectService = require('../../entities/project/project.service');
+const ProjectService = require('../../entities/project/project.service');
+const PayloadGeneratorService = require('../../common/services/payload-generator.service');
 
-project.get('/:id', (req, res) => {
-	projectService.findById(Number(req.params.id), (err, data) => {
-		if (!err) {
-			res.json(data);
-		} else {
-			res.status(400);
-			res.end();
-		}
-	});
-});
-
-project.post('/', (req, res) => {
-	projectService.save(req.body, (err, data) => {
-		if (!err) {
-			res.json(data);
-		} else {
-			res.status(400);
-			res.end();
-		}
-	});
-});
-
-project.put('/:id', (req, res) => {
-	projectService.update(Number(req.params.id), req.body, (err, data) => {
-		if (!err) {
-			res.json(data);
-		} else {
-			res.status(400);
-			res.end();
-		}
-	});
-});
-
-project.delete('/:id', (req, res) => {
-	projectService.removeById(Number(req.params.id), (err, data) => {
-		if (!err) {
-			res.json(data);
-		} else {
-			res.status(400);
-			res.end();
-		}
-	});
+project.get('/', (req, res, next) => {
+	ProjectService.getAll()
+		.then(PayloadGeneratorService.nextWithData(next, res))
+		.catch(next);
 });
 
 module.exports = project;
