@@ -47,26 +47,22 @@ class UserService {
 					},
 					(user, callback) => {
 						this.UserRepository.save(user)
-							.then(data =>
-								callback(null, {
-									tokenSecret: TokenService.createToken(data),
-									user: this.createUserPayload(
-										data.dataValues
-									)
-								})
-							)
+							.then(data => callback(null, {
+								tokenSecret: TokenService.createToken(data),
+								user: this.createUserPayload(
+									data.dataValues
+								)
+							}))
 							.catch(err => callback(err, null));
 					},
 					(payload, callback) => {
 						CompanyService.saveCompany(obj)
-							.then(data =>
-								callback(
-									null,
-									Object.assign({}, payload, {
-										company: data.dataValues
-									})
-								)
-							)
+							.then(data => callback(
+								null,
+								Object.assign({}, payload, {
+									company: data.dataValues
+								})
+							))
 							.catch(err => {
 								callback(err, null);
 							});
@@ -76,14 +72,12 @@ class UserService {
 							payload.user.id,
 							payload.company.id
 						)
-							.then(data =>
-								callback(
-									null,
-									Object.assign({}, payload, {
-										companyUser: data.dataValues
-									})
-								)
-							)
+							.then(data => callback(
+								null,
+								Object.assign({}, payload, {
+									companyUser: data.dataValues
+								})
+							))
 							.catch(err => {
 								callback(err, null);
 							});
@@ -93,14 +87,12 @@ class UserService {
 							name: 'General',
 							companyId: payload.company.id
 						})
-							.then(data =>
-								callback(
-									null,
-									Object.assign({}, payload, {
-										group: data.dataValues
-									})
-								)
-							)
+							.then(data => callback(
+								null,
+								Object.assign({}, payload, {
+									group: data.dataValues
+								})
+							))
 							.catch(err => callback(err, null));
 					},
 					(payload, callback) => {
@@ -108,14 +100,12 @@ class UserService {
 							payload.user.id,
 							payload.group.id
 						)
-							.then(data =>
-								callback(
-									null,
-									Object.assign({}, payload, {
-										groupUser: data.dataValues
-									})
-								)
-							)
+							.then(data => callback(
+								null,
+								Object.assign({}, payload, {
+									groupUser: data.dataValues
+								})
+							))
 							.catch(err => callback(err, null));
 					}
 				],
@@ -268,8 +258,7 @@ class UserService {
 	createUserPayload(data) {
 		this.payload = {
 			id: data.id,
-			firstName: data.firstName,
-			lastName: data.lastName,
+			name: data.name,
 			email: data.email
 		};
 		return this.payload;
