@@ -1,13 +1,13 @@
 const file = require('express').Router();
 const fileUpload = require('express-fileupload');
-const fileService = require('../common/services/file.services/file.service');
 const PayloadGeneratorService = require('../common/services/payload-generator.service');
+const parseJSAsync = require('../common/services/worker.service');
 
 // middleware for uploading files
 file.use(fileUpload());
 
 file.post('/parse', (req, res, next) => {
-	fileService(req.files, req.body.text, req.body.link)
+	parseJSAsync(req.files, req.body.text, req.body.link)
 		.then(PayloadGeneratorService.nextWithData(next, res))
 		.catch(next);
 });
